garbage input handling (i tried, but cpp is a hella tough enemy, so i've done my best) (requires refactoring)

+++ Rigidbody.cpp
@@ -28,7 +28,7 @@ namespace rocket_lab {
 		delete fieldHeight;
 	}
 
-	void Rigidbody::runPhysics(float dt) 
+	void Rigidbody::runPhysics(double dt) 
 	{
 		float totalAccX = accX, totalAccY = accY;
